RMS Value

+++ Program2.py
@@ -5,7 +5,7 @@ value = input('Enter the values')
 mylist = value.split(",")
 n = len(mylist)
 for i in range (n):
-    mylist[i] = int(mylist[i])
+    mylist[i] = float(mylist[i])
 for i in range (n):
     mylist[i] *= mylist[i]
